Add probe functionality to modprobe
Often times you want to use modprobe to load a module if a hardware
device is present. Adding `state=probe` that will attempt to load a
module and report changed if the module was loaded, no change if the
hardware device was not present and error if another error occurred.
This is the same behavior as 'present' except no error is raised or
change is reported if the driver reports no such device (ENODEV).

Previously submitted to ansible: ansible/ansible#6959